Deduplicate util
